Move 59 pure unit tests back to inline #[cfg(test)] modules
Addresses Devin PR review feedback: pure unit tests that don't need
a git repository were unnecessarily moved to integration test files
during the TmpRepo migration. Moving them back to inline modules
reduces integration test compile/link overhead.

Affected modules: post_commit, stats, ci_context, ignore, repository,
refs, prompt_utils.
