+++ src/git/repository.rs
@@ -3549,3 +3549,213 @@ fn parse_hunk_header(line: &str) -> Option<(Vec<u32>, bool, u32)> {
 
     Some((lines, is_pure_insertion, old_count))
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_parse_git_version_standard() {
+        // Standard git version format
+        assert_eq!(parse_git_version("git version 2.39.3"), Some((2, 39, 3)));
+        assert_eq!(parse_git_version("git version 2.23.0"), Some((2, 23, 0)));
+        assert_eq!(parse_git_version("git version 1.8.5"), Some((1, 8, 5)));
+    }
+
+    #[test]
+    fn test_parse_git_version_apple_git() {
+        // macOS Apple Git format
+        assert_eq!(
+            parse_git_version("git version 2.39.3 (Apple Git-146)"),
+            Some((2, 39, 3))
+        );
+    }
+
+    #[test]
+    fn test_parse_git_version_windows() {
+        // Windows git format
+        assert_eq!(
+            parse_git_version("git version 2.42.0.windows.2"),
+            Some((2, 42, 0))
+        );
+    }
+
+    #[test]
+    fn test_parse_git_version_no_patch() {
+        // Version without patch number
+        assert_eq!(parse_git_version("git version 2.39"), Some((2, 39, 0)));
+    }
+
+    #[test]
+    fn test_parse_git_version_with_newline() {
+        // Version string with trailing newline
+        assert_eq!(parse_git_version("git version 2.39.3\n"), Some((2, 39, 3)));
+    }
+
+    #[test]
+    fn test_parse_git_version_invalid() {
+        // Invalid formats should return None
+        assert_eq!(parse_git_version(""), None);
+        assert_eq!(parse_git_version("not a version"), None);
+        assert_eq!(parse_git_version("git version"), None);
+        assert_eq!(parse_git_version("git version x.y.z"), None);
+    }
+
+    #[test]
+    fn disable_internal_git_hooks_guard_applies_to_spawned_threads() {
+        let args = vec!["status".to_string()];
+        let _guard = disable_internal_git_hooks();
+
+        let spawned_args = args.clone();
+        let forwarded =
+            std::thread::spawn(move || args_with_disabled_hooks_if_needed(&spawned_args))
+                .join()
+                .expect("thread should join");
+
+        assert_eq!(forwarded[0], "-c");
+        assert!(forwarded[1].starts_with("core.hooksPath="));
+    }
+
+    #[test]
+    fn patch_profile_applies_canonical_machine_parse_flags() {
+        let args = vec!["diff".to_string(), "HEAD^".to_string(), "HEAD".to_string()];
+        let rewritten = args_with_internal_git_profile(&args, InternalGitProfile::PatchParse);
+
+        assert!(rewritten.iter().any(|arg| arg == "--no-ext-diff"));
+        assert!(rewritten.iter().any(|arg| arg == "--no-textconv"));
+        assert!(rewritten.iter().any(|arg| arg == "--src-prefix=a/"));
+        assert!(rewritten.iter().any(|arg| arg == "--dst-prefix=b/"));
+        assert!(rewritten.iter().any(|arg| arg == "--no-relative"));
+        assert!(rewritten.iter().any(|arg| arg == "--no-color"));
+        assert!(
+            rewritten
+                .iter()
+                .any(|arg| arg == "--diff-algorithm=default")
+        );
+        assert!(rewritten.iter().any(|arg| arg == "--indent-heuristic"));
+        assert!(rewritten.iter().any(|arg| arg == "--inter-hunk-context=0"));
+    }
+
+    #[test]
+    fn numstat_profile_disables_renames_and_external_renderers() {
+        let args = vec![
+            "diff".to_string(),
+            "--numstat".to_string(),
+            "HEAD^".to_string(),
+            "HEAD".to_string(),
+        ];
+        let rewritten = args_with_internal_git_profile(&args, InternalGitProfile::NumstatParse);
+        assert!(rewritten.iter().any(|arg| arg == "--no-ext-diff"));
+        assert!(rewritten.iter().any(|arg| arg == "--no-textconv"));
+        assert!(rewritten.iter().any(|arg| arg == "--no-color"));
+        assert!(rewritten.iter().any(|arg| arg == "--no-relative"));
+        assert!(rewritten.iter().any(|arg| arg == "--no-renames"));
+    }
+
+    #[test]
+    fn numstat_profile_strips_short_rename_and_copy_flags() {
+        let args = vec![
+            "diff".to_string(),
+            "--numstat".to_string(),
+            "-M90%".to_string(),
+            "-C".to_string(),
+            "-C75%".to_string(),
+            "HEAD^".to_string(),
+            "HEAD".to_string(),
+        ];
+        let rewritten = args_with_internal_git_profile(&args, InternalGitProfile::NumstatParse);
+        assert!(!rewritten.iter().any(|arg| arg == "-C"));
+        assert!(!rewritten.iter().any(|arg| arg.starts_with("-M")));
+        assert!(!rewritten.iter().any(|arg| arg.starts_with("-C")));
+        assert!(rewritten.iter().any(|arg| arg == "--no-renames"));
+    }
+
+    #[test]
+    fn general_profile_is_noop() {
+        let args = vec!["status".to_string(), "--porcelain=v2".to_string()];
+        let rewritten = args_with_internal_git_profile(&args, InternalGitProfile::General);
+        assert_eq!(rewritten, args);
+    }
+
+    #[test]
+    fn patch_profile_strips_conflicting_ext_diff_and_color_flags() {
+        let args = vec![
+            "diff".to_string(),
+            "--ext-diff".to_string(),
+            "--color=always".to_string(),
+            "HEAD".to_string(),
+        ];
+        let rewritten = args_with_internal_git_profile(&args, InternalGitProfile::PatchParse);
+
+        assert!(rewritten.iter().any(|arg| arg == "--no-ext-diff"));
+        assert!(!rewritten.iter().any(|arg| arg == "--ext-diff"));
+        assert!(!rewritten.iter().any(|arg| arg.starts_with("--color")));
+        assert!(rewritten.iter().any(|arg| arg == "--no-color"));
+    }
+
+    #[test]
+    fn patch_profile_strips_split_prefix_args() {
+        let args = vec![
+            "diff".to_string(),
+            "--src-prefix".to_string(),
+            "SRC/".to_string(),
+            "--dst-prefix".to_string(),
+            "DST/".to_string(),
+            "HEAD^".to_string(),
+            "HEAD".to_string(),
+        ];
+        let rewritten = args_with_internal_git_profile(&args, InternalGitProfile::PatchParse);
+
+        assert!(!rewritten.iter().any(|arg| arg == "--src-prefix"));
+        assert!(!rewritten.iter().any(|arg| arg == "--dst-prefix"));
+        assert!(!rewritten.iter().any(|arg| arg == "SRC/"));
+        assert!(!rewritten.iter().any(|arg| arg == "DST/"));
+        assert!(rewritten.iter().any(|arg| arg == "--src-prefix=a/"));
+        assert!(rewritten.iter().any(|arg| arg == "--dst-prefix=b/"));
+    }
+
+    #[test]
+    fn profile_rewrite_does_not_strip_pathspec_tokens_after_double_dash() {
+        let args = vec![
+            "diff".to_string(),
+            "--color=always".to_string(),
+            "HEAD^".to_string(),
+            "HEAD".to_string(),
+            "--".to_string(),
+            "--color".to_string(),
+            "--relative".to_string(),
+            "file.txt".to_string(),
+        ];
+        let rewritten = args_with_internal_git_profile(&args, InternalGitProfile::PatchParse);
+        let separator = rewritten
+            .iter()
+            .position(|arg| arg == "--")
+            .expect("rewritten args should keep pathspec separator");
+        assert_eq!(
+            rewritten[separator + 1..],
+            [
+                "--color".to_string(),
+                "--relative".to_string(),
+                "file.txt".to_string()
+            ]
+        );
+    }
+
+    #[test]
+    fn raw_diff_profile_keeps_rename_flags_untouched() {
+        let args = vec![
+            "diff".to_string(),
+            "--raw".to_string(),
+            "-z".to_string(),
+            "-M".to_string(),
+            "HEAD^".to_string(),
+            "HEAD".to_string(),
+        ];
+        let rewritten = args_with_internal_git_profile(&args, InternalGitProfile::RawDiffParse);
+        assert!(rewritten.iter().any(|arg| arg == "-M"));
+        assert!(rewritten.iter().any(|arg| arg == "--no-ext-diff"));
+        assert!(rewritten.iter().any(|arg| arg == "--no-textconv"));
+        assert!(rewritten.iter().any(|arg| arg == "--no-color"));
+        assert!(rewritten.iter().any(|arg| arg == "--no-relative"));
+    }
+}
