+++ tests/integration/ci_context_unit.rs
@@ -1,38 +1,7 @@
 use crate::repos::test_repo::TestRepo;
-use git_ai::ci::ci_context::{CiContext, CiEvent, CiRunResult};
+use git_ai::ci::ci_context::{CiContext, CiEvent};
 use git_ai::git::repository::find_repository_in_path;
 use std::fs;
-
-#[test]
-fn test_ci_event_debug() {
-    let event = CiEvent::Merge {
-        merge_commit_sha: "abc123".to_string(),
-        head_ref: "feature".to_string(),
-        head_sha: "def456".to_string(),
-        base_ref: "main".to_string(),
-        base_sha: "ghi789".to_string(),
-    };
-
-    let debug_str = format!("{:?}", event);
-    assert!(debug_str.contains("Merge"));
-    assert!(debug_str.contains("abc123"));
-    assert!(debug_str.contains("feature"));
-}
-
-#[test]
-fn test_ci_run_result_debug() {
-    let result = CiRunResult::SkippedSimpleMerge;
-    let debug_str = format!("{:?}", result);
-    assert!(debug_str.contains("SkippedSimpleMerge"));
-
-    let result2 = CiRunResult::SkippedFastForward;
-    let debug_str2 = format!("{:?}", result2);
-    assert!(debug_str2.contains("SkippedFastForward"));
-
-    let result3 = CiRunResult::NoAuthorshipAvailable;
-    let debug_str3 = format!("{:?}", result3);
-    assert!(debug_str3.contains("NoAuthorshipAvailable"));
-}
 
 #[test]
 fn test_ci_context_with_repository() {
